requirements.txt
# HerHealthAI - Complete Requirements
# Core API
fastapi
uvicorn[standard]

# ML & Data
pandas
numpy
scikit-learn
joblib
openpyxl
xgboost
shap

# Frontend
streamlit

# Database
sqlalchemy

# Auth & Security
PyJWT
passlib[bcrypt]
python-dotenv

# LLM
google-genai

# HTTP
requests
httpx
slowapi

# Testing
pytest
pytest-asyncio
schedule
